0521 1차

+++ src/main/java/kr/blug/tour/controller/LikesContentController.java
@@ -38,33 +38,72 @@ public class LikesContentController {
 
 		        SaveResponseDto result = likesContentService.saveLikesContent(dto);
 
-		        return ResponseEntity.ok(Map.of("result", result));
+		        if(result.isSuccess()) {
+		        	return ResponseEntity.ok(Map.of(
+		        			"result", "success", 
+		        			"message", result.getMessage(), 
+		        			"id_name", result.getId_name(), 
+		        			"value", result.getId() ));
+		        } 
+		        else {
+		        	return ResponseEntity.ok(Map.of(
+		        			"result", result.getMessage()));
+		        }
 	}
 	
 	@DeleteMapping("/likes/content/delete")
 	public ResponseEntity<Map<String, Object>> deleteLikesContent(
-				@RequestParam(name="user_id", required = true) Long userId,
-				@RequestParam(name="contentid", required = true)  String contentId
+				@RequestParam(name="user_id", required = false) Long userId,
+				@RequestParam(name="contentid", required = false)  String contentId
 			) {
+		
+		if(userId == null) {
+			return ResponseEntity.ok(Map.of("result", "error - parameter user_id is required"));				
+		}
+		
+		if(contentId == null) {
+			return ResponseEntity.ok(Map.of("result", "error - parameter contentId is required"));	
+		}
 		
 		SaveResponseDto result = likesContentService.deleteByUserIdAndContentId(userId, contentId);
 		
-		return ResponseEntity.ok(Map.of("result", result));
+        if(result.isSuccess()) {
+        	return ResponseEntity.ok(Map.of(
+        			"result", "success", 
+        			"message", result.getMessage(), 
+        			"id_name", result.getId_name(), 
+        			"value", result.getId() ));
+        } 
+        else {
+        	return ResponseEntity.ok(Map.of(
+        			"result", result.getMessage()));
+        }
 	}
 	
 	
 	@GetMapping("/likes/content/check")
 	public ResponseEntity<Map<String, Object>> findByUserAndContent(
-				@RequestParam("user_id") Long userId,
-				@RequestParam("contentid") String contentId){
+				@RequestParam(name = "user_id", required = false) Long userId,
+				@RequestParam(name= "contentid", required = false) String contentId){
+		
+		if(userId == null) {
+			return ResponseEntity.ok(Map.of("result", "error - parameter user_id is required"));				
+		}
+		
+		if(contentId == null) {
+			return ResponseEntity.ok(Map.of("result", "error - parameter contentId is required"));	
+		}
 		
 		Optional<LikesContentDto> dto = likesContentService.findByUserAndContent(userId, contentId); 
 		
 		if(dto.isPresent()) {
-			return ResponseEntity.ok(Map.of("resulte", "success", "item", dto));
+			return ResponseEntity.ok(Map.of(
+					"resulte", "success", 
+					"item", dto));
 		}
 		else {
-			return ResponseEntity.ok(Map.of("result", "not_found"));
+			return ResponseEntity.ok(Map.of(
+					"result", "not_found"));
 		}
 	}
 	
@@ -100,7 +139,8 @@ public class LikesContentController {
 	public ResponseEntity<Map<String, Object>> listLikesContentAll(
 				@RequestParam(name="areacode", required = false) String areaCode,
 				@RequestParam(name="sigungucode", required = false) String sigunguCode,
-				@RequestParam(name="user_id", required = false) Long userId,				
+				@RequestParam(name="user_id", required = false) Long userId,			
+				@RequestParam(name="contenttypeid", required = false) String contentTypeId,
 				@PageableDefault(size=10, page=0) Pageable pageable
 			) {
 		
@@ -108,11 +148,16 @@ public class LikesContentController {
 //		if(sigunguCode == null) sigunguCode  = "%";
 //		if(userId == null) userId = "%";
 		
-		Page<LikesContentDto> items = likesContentService.listLikesContentAll(pageable, areaCode, sigunguCode, userId);
+		System.out.println("contentTypeId = " + contentTypeId);
+		
+		
+		
+		Page<LikesContentDto> items = likesContentService.listLikesContentAll(pageable, areaCode, sigunguCode, userId, contentTypeId);
 		if(!items.isEmpty()) {
 		
 			return ResponseEntity.ok(Map.of(
-					"result", "success", "items", items,
+					"result", "success", 
+					"items", items,
 					"totalPages", items.getTotalPages(),
 					"totalElements", items.getTotalElements(),				
 					"currentPage", items.getNumber()
